refactor: 상수 파일로 묶기 및 상수 파일 소문자로 이름 변경
