+++ src/hooks/useIsEmojiSendable.ts
@@ -3,7 +3,7 @@
 import { useChatInfo } from '@/store/chatInfo';
 import { useAgora } from '@/store/agora';
 import { useShallow } from 'zustand/react/shallow';
-import { AGORA_STATUS } from '@/constants/Agora';
+import { AGORA_STATUS } from '@/constants/agora';
 
 export default function useIsEmojiSendable() {
   // end 상태가 있으면 이모지 전송 불가
